Move MySQLQuery to autoloading class
